% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fertdata.R
\name{clean_fertdata}
\alias{clean_fertdata}
\title{cleans fertility data; not exported (for internal use only)}
\usage{
clean_fertdata(raw_fert_path)
}
\arguments{
\item{raw_fert_path}{path to raw mortality data}
}
\description{
cleans fertility data; not exported (for internal use only)
}
\examples{
clean_fertdata("data-raw/mort-data/malawi.xlsx")
}
