% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_idb_long.R
\name{clean_idb}
\alias{clean_idb}
\title{Pivots International Database exports from wide to long format}
\usage{
clean_idb(country, datadir, filetype = ".csv", skip = 0, is_legacy = FALSE)
}
\arguments{
\item{country}{Case sensitive name of file name (excepting file extension).}

\item{datadir}{Relative or absolute path.}

\item{filetype}{Default value is `.csv`}

\item{skip}{Default value is `1`}

\item{is_legacy}{Default value is `FALSE`}
}
\value{
A data frame.
}
\description{
This function is capable of cleaning two types of IDB formats.
}
\examples{
clean_idb("tanzania", "data-raw/", filetype = ".xlsx", is_legacy = TRUE)
clean_idb("kenya", "data-raw/", skip = 1)

}
\author{
Vania Wang
}
