% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mortdata.R
\name{clean_mortdata}
\alias{clean_mortdata}
\title{cleans mortality data; not exported (for internal use only)}
\usage{
clean_mortdata(raw_mort_path)
}
\arguments{
\item{raw_mort_path}{path to raw mortality data}
}
\description{
cleans mortality data; not exported (for internal use only)
}
\examples{
clean_mortdata("./data-raw/mort-data/malawi.xlsx")
}
