% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_popdata.R
\name{filter_popdata}
\alias{filter_popdata}
\title{filters an already clean pop dataset, according to params}
\usage{
filter_popdata(pop_data, yr = NULL, adm_level = 0, sex = c("B"))
}
\arguments{
\item{pop_data}{needs to be an already cleaned pop_data}

\item{yr}{a vec of values}

\item{adm_level}{an int}

\item{sex}{a vec of values}
}
\description{
filters an already clean pop dataset, according to params
}
