% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_to_shp.R
\name{join_data_to_shp}
\alias{join_data_to_shp}
\title{Title}
\usage{
join_data_to_shp(pop_data, pop_sf, adm_level = "ADM1", yr = 2020)
}
\arguments{
\item{adm_level}{A numeric specifying ADM level}

\item{yr}{}
}
\description{
Title
}
