% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_demodata_to_shp.R
\name{join_demodata_to_shp}
\alias{join_demodata_to_shp}
\title{Make all demographic data spatial}
\usage{
join_demodata_to_shp(demo_data, spatial_data, geo_match)
}
\arguments{
\item{demo_data}{df, must have a matching column named "area_name"}

\item{spatial_data}{sf, must have a matching column named "DIST_NAME"}
}
\value{
a df
}
\description{
Merges fertility and mortality data with country shapefiles at the
ADM2 level.
}
