% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_popdata_to_shp.R
\name{join_popdata_to_shp}
\alias{join_popdata_to_shp}
\title{Make Population Data Spatial}
\usage{
join_popdata_to_shp(pop_data, pop_sf, adm_level = "ADM1", yr = 2020)
}
\arguments{
\item{pop_data}{df}

\item{pop_sf}{sf}

\item{adm_level}{A numeric specifying ADM level}

\item{yr}{}
}
\value{
a joined dataframe
}
\description{
Make Population Data Spatial
}
