% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_idb_long.R
\name{pivot_idb_long}
\alias{pivot_idb_long}
\title{Pivots International Database exports from wide to long format}
\usage{
pivot_idb_long(
  country,
  datadir,
  filetype = ".csv",
  skip = 0,
  is_legacy = FALSE
)
}
\arguments{
\item{country}{Case sensitive name of file name (excepting file extension).}

\item{datadir}{Relative or absolute path to folder containing the input file.}

\item{filetype}{Default value is \code{.csv}.}

\item{skip}{Default value is \code{1}. An input for read_csv for skipping any
number of header rows.}

\item{is_legacy}{Default value is \code{FALSE} for processing the most recent
version of IDB files.}
}
\value{
A clean data frame with (possibly) the following columns: "ADM_LEVEL",
"GEO_MATCH", "ADM0", "ADM2", "YR", "SEX", "AGE_CAT", "POP_TOTAL"
}
\description{
This function is capable of pivoting two types of IDB formats.
The output format will have, at minimum, the following columns:
"ADM_LEVEL", "GEO_MATCH", "ADM0", "ADM2", "YR", "SEX", "AGE_CAT",
"POP_TOTAL"
}
\examples{
pivot_idb_long("tanzania", "data-raw/pop-data/", filetype = ".xlsx", is_legacy = TRUE)
pivot_idb_long("kenya", "data-raw/pop-data/", skip = 1)

}
\author{
Vania Wang
}
