% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fertbar.R
\name{plot_fertbar}
\alias{plot_fertbar}
\title{Creates total fertility rate visualization}
\usage{
plot_fertbar(fert_data, adm_level)
}
\arguments{
\item{fert_data}{Country mortality data in long format.}
}
\value{
A ggplot bar graph with total births per woman on the x axis and
district on the y-axis..
}
\description{
This function plots a bar graph showing total fertility rate.
}
\examples{
plot_mr(fert)

}
\author{
Britta Schumacher
}
