% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fertbar_adolescent.R
\name{plot_fertbar_adolescent}
\alias{plot_fertbar_adolescent}
\title{Creates adolescent fertility rate visualization}
\usage{
plot_fertbar_adolescent(fert_data, adm_level)
}
\arguments{
\item{fert_data}{Country mortality data in long format.}
}
\value{
A ggplot bar graph with total births per adolscent (ages 15-19)
woman on the x axis and district on the y-axis..
}
\description{
This function plots a bar graph showing adolescent fertility rate.
}
\examples{
plot_mr(fert)

}
\author{
Britta Schumacher
}
