% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lebar.R
\name{plot_lebar}
\alias{plot_lebar}
\title{Creates life expectancy visualization}
\usage{
plot_lebar(mort_data, adm_name)
}
\arguments{
\item{mort_data}{Country mortality data in long format.}

\item{adm_name}{Name of administrative geography (i.e., nation or district).}
}
\value{
A ggplot bar graph with life expectancy at birth on the x axis and sex (M/F)
on the y axis.
}
\description{
This function plots a bar graph showing life expectancy at birth
on the x axis and sex (M/F)on the y axis.
}
\examples{
plot_le(mort, "Malawi")

}
\author{
Britta Schumacher
}
