% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mortbar_infantchild.R
\name{plot_mortbar_infantchild}
\alias{plot_mortbar_infantchild}
\title{Plots infant and child mortality bar graphs}
\usage{
plot_mortbar_infantchild(mort_data, adm_name)
}
\arguments{
\item{mort_data}{Country mortality data in long format.}

\item{adm_name}{Name of administrative geography (i.e., nation or district).}
}
\value{
A ggplot bar graph with Deaths per 1,000 live births on the x axis and infant (< 1yr)
and child (< 5yrs) on the y axis.
}
\description{
This function plots a bar graph showing infant and child mortality.
}
\examples{
plot_mr(mort_data, "Malawi")

}
\author{
Britta Schumacher
}
