% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop.R
\name{plot_pop}
\alias{plot_pop}
\title{Creates population by District visualization}
\usage{
plot_pop(pop_data, adm_name)
}
\arguments{
\item{pop_data}{Country demographic data in long format.}
}
\value{
A ggplot bar graph with total population on the x axis and
district on the y-axis. Plot highlights specific geography when "Malawi"
is not selected in the geography dropdown.
}
\description{
This function plots a bar graph showing the population breakdown
by district organized from greatest to least population in 2018.
}
\examples{
plot_pop(pop,"Balaka")

}
\author{
Britta Schumacher
}
