% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyr.R
\name{plot_pyr}
\alias{plot_pyr}
\title{required data format ---
selected year: e.g. 2015
adm selection: e.g. 1
sex: only M/F}
\usage{
plot_pyr(pop_data, adm_name, age_group)
}
\description{
required data format ---
selected year: e.g. 2015
adm selection: e.g. 1
sex: only M/F
}
