% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyr.R
\name{plot_pyr_subnat}
\alias{plot_pyr_subnat}
\title{Plot subnational population pyramids}
\usage{
plot_pyr_subnat(country_data, age_group, adm_var = "ADM0")
}
\arguments{
\item{country_data}{A data frame containing population data.}

\item{age_group}{A choice between \code{one} or \code{five}.}

\item{adm_var}{The column containing the ADM level of each entry. Defaults to \code{ADM_LEVEL}.}
}
\value{
a list of plots
}
\description{
\code{plot_pyr_subnat} generates a list population pyramid plots. The input data is
\code{pop_data} and has strict requirements for the pyramid to render correctly:
}
\details{
\itemize{
\item The YR column should have one year only
\item The SEX column should have two values only
}
}
