% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sex.R
\name{plot_sex}
\alias{plot_sex}
\title{Creates population by sex visualization}
\usage{
plot_sex(pop_data, adm_name)
}
\arguments{
\item{pop_data}{Country demographic data in long format.}
}
\value{
A ggplot bar graph with total population on the x axis and
male / female on the y-axis..
}
\description{
This function plots a bar graph showing the population breakdown by sex in a given geography.
}
\examples{
plot_sex(pop,"Malawi")

}
\author{
Britta Schumacher
}
