% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_urbrur.R
\name{plot_urbrur}
\alias{plot_urbrur}
\title{Creates urban / rural population visualization}
\usage{
plot_urbrur(pop_data, adm_name)
}
\arguments{
\item{pop_data}{Country demographic data in long format.}
}
\value{
A ggplot bar graph with total population on the x axis and
urban / rural on the y-axis..
}
\description{
This function plots a bar graph showing the urban / rural population breakdown in a given geography.
}
\examples{
plot_urban_rural(pop,"Malawi")

}
\author{
Britta Schumacher
}
