% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{pretty_agecats}
\alias{pretty_agecats}
\title{Make five-year age categories pretty.}
\usage{
pretty_agecats(agecat_vec)
}
\arguments{
\item{agecat_vec}{A vector of strings specifying age categories.}
}
\value{
A pretty vector of strings
}
\description{
pop_data five-year age categories are formatted as "0_4",
"5_9", ... , "75_79", "80_". This function reformats these categories as
"0-4", "5-9", ... , "75-79", "80+".
}
