% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{pretty_symmetric}
\alias{pretty_symmetric}
\title{Make scale symmetric and pretty}
\usage{
pretty_symmetric(vec)
}
\arguments{
\item{vec}{An ascending sequence of numbers greater than 0.}
}
\value{
A pretty numeric vector
}
\description{
Developed specifically for plotting population pyramids. Given
a ascending sequence of numbers, creates a symmetric scale mirrored across
the number line, with 0 as the midpoint.
}
