% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{theme_plot}
\alias{theme_plot}
\title{Defines the plot themes for the mortality-related plot functions.}
\usage{
theme_plot(...)
}
\description{
This function charts the theme for the mortality plots in OSDS.
}
\author{
Britta Schumacher
}
