% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_popdata.R
\name{validate_popdata}
\alias{validate_popdata}
\title{Validates the imported population data}
\usage{
validate_popdata(popdf)
}
\arguments{
\item{popdf}{}
}
\description{
Validates the imported population data
}
