% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_popdata.R
\name{validate_popdata}
\alias{validate_popdata}
\title{Validate the imported population data}
\usage{
validate_popdata(pop_data)
}
\arguments{
\item{pop_data}{A data frame containing population data}
}
\value{
A logical
}
\description{
Validate the imported population data
}
